backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
weaviate-client==4.6.5
sentence-transformers==3.0.1
python-dotenv==1.0.1
ollama==0.3.3
tiktoken==0.8.0
pydantic==2.9.2
python-multipart==0.0.12
requests==2.32.3
PyMuPDF==1.24.13
